refactor(server/servicebroker): clean up handlers
