+++ pkg/server/servicebroker/apiserver_test.go
@@ -1,4 +1,4 @@
-package broker
+package servicebroker
 
 import (
 	"testing"
